PS2 PriceResponse interoperable, fix type primary key in JPA, simplify tests

+++ src/main/java/com/price/priceservice/repository/PriceRepository.java
@@ -10,7 +10,7 @@ import java.time.LocalDateTime;
 import java.util.List;
 
 @Repository
-public interface PriceRepository extends JpaRepository<Price, Long> {
+public interface PriceRepository extends JpaRepository<Price, BigInteger> {
     
     List<Price> findByProductIdAndBrandIdAndStartDateLessThanEqualAndEndDateGreaterThanEqualOrderByPriorityDesc(
             BigInteger productId, Integer brandId, LocalDateTime startDate, LocalDateTime endDate);
