+++ src/main/java/com/price/priceservice/service/response/PriceResponse.java
@@ -7,8 +7,6 @@ import lombok.NoArgsConstructor;
 import lombok.ToString;
 
 import java.io.Serializable;
-import java.math.BigDecimal;
-import java.math.BigInteger;
 
 @Getter
 @Setter
@@ -19,11 +17,11 @@ public class PriceResponse implements Serializable {
 
     private static final long serialVersionUID = 1L;
 
-    private Integer brandId;
-    private BigInteger productId;
-    private Integer priceList;
+    private int brandId;
+    private long productId;
+    private int priceList;
     private String startDate;
     private String endDate;
-    private BigDecimal price;
+    private float price;
 
 }
